acse_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
redis==5.0.4
xgboost==2.0.3
scikit-learn==1.4.2
shap==0.45.1
numpy==1.26.4
python-jose[cryptography]==3.3.0
httpx==0.27.0          # for tests
pytest==8.2.0
pytest-asyncio==0.23.6
